+++ tests/testthat/test-combining.R
@@ -1,6 +1,9 @@
 test_that("unifont_combining()", {
-    expect_length(unifont_combining(), 178L)
-    expect_length(unifont_combining(upper = FALSE), 178L)
+    expect_equal(unifont_version(),
+                 numeric_version("15.1.04"))
+
+    expect_length(unifont_combining(), 184L)
+    expect_length(unifont_combining(upper = FALSE), 184L)
     expect_length(unifont_combining(csur = FALSE), 0L)
-    expect_length(unifont_combining(unicode = TRUE), 2350L)
+    expect_length(unifont_combining(unicode = TRUE), 2356L)
 })